feat: separate writer API in different header
